Support specifying a firewall group when updating a rule
This fixes a bug that prevents updating a firewall rule when it has a
specified but unchanged group property, and some other property requires
an update. In this case the group property must be excluded from the
call to Set-NetFirewallRule in Set-TargetResource, as it cannot be used
in conjunction with Name.